Change project module structure and Change test file name
